Create Code
This code loads an image file using the PIL library, generates a secret key, pads the image data to match the block size of AES, encrypts the data using the CBC mode of AES, and then saves the encrypted data to a file. Note that this is a simplified example and may not provide sufficient security for all use cases.